@@ -10,7 +10,7 @@ abstract class BaseWebhookController
     /**
      * @var \Laravel\Cashier\Mollie\Contracts\GetMolliePayment
      */
-    protected $getMolliePayment;
+    protected GetMolliePayment $getMolliePayment;
 
     public function __construct(GetMolliePayment $getMolliePayment)
     {
@@ -21,14 +21,15 @@ abstract class BaseWebhookController
      * Fetch a payment from Mollie using its ID.
      * Returns null if the payment cannot be retrieved.
      *
-     * @param $id
+     * @param string $id
+     * @param array $parameters
      * @return \Mollie\Api\Resources\Payment|null
      * @throws \Mollie\Api\Exceptions\ApiException
      */
-    public function getPaymentById($id)
+    public function getMolliePaymentById(string $id, array $parameters = [])
     {
         try {
-            return $this->getMolliePayment->execute($id);
+            return $this->getMolliePayment->execute($id, $parameters);
         } catch (ApiException $e) {
             if (! config('app.debug')) {
                 // Prevent leaking information
