@@ -17,12 +17,35 @@ use Laravel\Cashier\Events\OrderProcessed;
 use Laravel\Cashier\Exceptions\InvalidMandateException;
 use Laravel\Cashier\MandatedPayment\MandatedPaymentBuilder;
 use Laravel\Cashier\Order\Contracts\MinimumPayment;
+use Laravel\Cashier\Payment;
+use Laravel\Cashier\Refunds\Refund;
+use Laravel\Cashier\Refunds\RefundBuilder;
 use Laravel\Cashier\Traits\HasOwner;
 use LogicException;
 use Mollie\Api\Resources\Mandate;
+use Mollie\Api\Resources\Payment as MolliePayment;
 use Mollie\Api\Types\PaymentStatus;
 
 /**
+ * @property int id
+ * @property string owner_type
+ * @property int owner_id
+ * @property string number
+ * @property string currency
+ * @property int subtotal
+ * @property int tax
+ * @property int total
+ * @property int balance_before
+ * @property int credit_used
+ * @property int total_due
+ * @property string mollie_payment_id
+ * @property string mollie_payment_status
+ * @property \Carbon\Carbon|null processed_at
+ * @property int amount_refunded
+ * @property int amount_charged_back
+ * @property \Laravel\Cashier\Order\OrderItemCollection items
+ * @property \Laravel\Cashier\Refunds\RefundCollection refunds
+ * @property \Laravel\Cashier\Payment payment;
  * @method static create(array $data)
  */
 class Order extends Model
@@ -226,6 +249,16 @@ class Order extends Model
     }
 
     /**
+     * The refunds for this order.
+     *
+     * @return HasMany
+     */
+    public function refunds()
+    {
+        return $this->hasMany(Refund::class);
+    }
+
+    /**
      * Create a new Eloquent Collection instance.
      *
      * @param  array  $models
@@ -343,7 +376,7 @@ class Order extends Model
      * @param $id
      * @return self
      */
-    public static function findByPaymentId($id)
+    public static function findByMolliePaymentId($id)
     {
         return self::where('mollie_payment_id', $id)->first();
     }
@@ -355,7 +388,7 @@ class Order extends Model
      * @return self
      * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
      */
-    public static function findByPaymentIdOrFail($id)
+    public static function findByMolliePaymentIdOrFail($id)
     {
         return self::where('mollie_payment_id', $id)->firstOrFail();
     }
@@ -376,11 +409,12 @@ class Order extends Model
      * Restores any credit used to the customer's balance and resets the credits applied to the Order.
      * Invokes handlePaymentFailed() on each related OrderItem.
      *
+     * @param \Mollie\Api\Resources\Payment $molliePayment
      * @return $this
      */
-    public function handlePaymentFailed()
+    public function handlePaymentFailed(MolliePayment $molliePayment)
     {
-        return DB::transaction(function () {
+        return DB::transaction(function () use ($molliePayment) {
             if ($this->creditApplied()) {
                 $this->owner->addCredit($this->getCreditUsed());
             }
@@ -389,6 +423,11 @@ class Order extends Model
                 'mollie_payment_status' => 'failed',
                 'balance_before' => 0,
                 'credit_used' => 0,
+            ]);
+
+            $localPayment = Payment::findByPaymentIdOrFail($molliePayment->id);
+            $localPayment->update([
+                'mollie_payment_status' => 'failed',
             ]);
 
             Event::dispatch(new OrderPaymentFailed($this));
@@ -407,12 +446,19 @@ class Order extends Model
      * Handles a paid payment for this order.
      * Invokes handlePaymentPaid() on each related OrderItem.
      *
+     * @param \Mollie\Api\Resources\Payment $molliePayment
      * @return $this
      */
-    public function handlePaymentPaid()
+    public function handlePaymentPaid(MolliePayment $molliePayment)
     {
-        return DB::transaction(function () {
+        return DB::transaction(function () use ($molliePayment) {
             $this->update(['mollie_payment_status' => 'paid']);
+
+            $localPayment = Payment::findByPaymentIdOrFail($molliePayment->id);
+            $localPayment->update([
+                'mollie_payment_status' => 'paid',
+            ]);
+
             Event::dispatch(new OrderPaymentPaid($this));
 
             $this->items->each(function (OrderItem $item) {
@@ -480,11 +526,57 @@ class Order extends Model
     }
 
     /**
+     * @return \Money\Money
+     */
+    public function getAmountRefunded()
+    {
+        return $this->toMoney($this->amount_refunded);
+    }
+
+    /**
+     * @return \Money\Money
+     */
+    public function getAmountChargedBack()
+    {
+        return $this->toMoney($this->amount_charged_back);
+    }
+
+    /**
      * @return string
      */
     public function getCurrency()
     {
         return $this->currency;
+    }
+
+    /**
+     * Get an empty refund builder for this order.
+     *
+     * @return \Laravel\Cashier\Refunds\RefundBuilder
+     */
+    public function refundBuilder()
+    {
+        return RefundBuilder::forOrder($this);
+    }
+
+    /**
+     * Get a refund builder prepared to completely refund this order.
+     *
+     * @return \Laravel\Cashier\Refunds\RefundBuilder
+     */
+    public function completeRefundBuilder()
+    {
+        return RefundBuilder::forWholeOrder($this);
+    }
+
+    /**
+     * Initiate a complete refund for this order.
+     *
+     * @return \Laravel\Cashier\Refunds\Refund
+     */
+    public function refundCompletely()
+    {
+        return $this->completeRefundBuilder()->create();
     }
 
     /**
@@ -522,5 +614,15 @@ class Order extends Model
         }
 
         return $minimumPaymentAmount;
+    }
+
+    /**
+     * The payments for this order.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function payments(): HasMany
+    {
+        return $this->hasMany(Payment::class)->orderByDesc('updated_at');
     }
 }
