@@ -12,6 +12,7 @@ use Laravel\Cashier\FirstPayment\FirstPaymentBuilder;
 use Laravel\Cashier\Plan\Contracts\PlanRepository;
 use Laravel\Cashier\Plan\Plan;
 use Laravel\Cashier\SubscriptionBuilder\Contracts\SubscriptionBuilder as Contract;
+use Money\Money;
 
 /**
  * Creates and configures a Mollie first payment to create a new mandate via Mollie's checkout
@@ -89,18 +90,20 @@ class FirstPaymentSubscriptionBuilder implements Contract
         if ($this->isTrial) {
             $taxPercentage = $this->owner->taxPercentage() * 0.01;
             $total = $this->plan->firstPaymentAmount();
-
+          
             if ($total->isZero()) {
                 $vat = $total->subtract($total); // zero VAT
             } else {
-                $vat = $total->divide(1 + $taxPercentage)->multiply($taxPercentage);
+                $vat = $total->divide(1 + $taxPercentage)
+                             ->multiply($taxPercentage, $this->roundingMode($total, $taxPercentage));
             }
             $subtotal = $total->subtract($vat);
 
             $actions[] = new AddGenericOrderItem(
                 $this->owner,
                 $subtotal,
-                $this->plan->firstPaymentDescription()
+                $this->plan->firstPaymentDescription(),
+                $this->roundingMode($total, $taxPercentage)
             );
         } elseif ($coupon) {
             $actions[] = new ApplySubscriptionCouponToPayment($this->owner, $coupon, $actions->processedOrderItems());
@@ -239,5 +242,33 @@ class FirstPaymentSubscriptionBuilder implements Contract
         $this->firstPaymentBuilder->setDescription($this->plan->firstPaymentDescription());
 
         return $this->firstPaymentBuilder;
+    }
+
+    /**
+     * Format the money as basic decimal
+     *
+     * @param \Money\Money $total
+     * @param float $taxPercentage
+     *
+     * @return int
+     */
+    public function roundingMode(Money $total, float $taxPercentage)
+    {
+        $vat = $total->divide(1 + $taxPercentage)->multiply($taxPercentage);
+
+        $subtotal = $total->subtract($vat);
+
+        $recalculatedTax = $subtotal->multiply($taxPercentage * 100)->divide(100);
+
+        $finalTotal = $subtotal->add($recalculatedTax);
+
+        if ($finalTotal->equals($total)) {
+            return Money::ROUND_HALF_UP;
+        }
+        if ($finalTotal->greaterThan($total)) {
+            return Money::ROUND_UP;
+        }
+
+        return  Money::ROUND_DOWN;
     }
 }
